+++ src/main/java/ru/rutmiit/Library/repositories/LibrarianRepository.java
@@ -3,5 +3,5 @@ package ru.rutmiit.Library.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.rutmiit.Library.entities.Librarian;
 
-public interface LibrarianRepository extends JpaRepository<Librarian, Long> {
+public interface LibrarianRepository extends JpaRepository<Librarian, Integer> {
 }
